move server haptics to dedicated server api
