+++ common/src/main/java/org/vivecraft/common/api_impl/VRAPIImpl.java
@@ -3,13 +3,11 @@ package org.vivecraft.common.api_impl;
 import net.minecraft.server.level.ServerPlayer;
 import net.minecraft.world.entity.player.Player;
 import org.vivecraft.api.VRAPI;
-import org.vivecraft.api.data.VRBodyPart;
 import org.vivecraft.api.data.VRPose;
 import org.vivecraft.api.data.VRPoseHistory;
 import org.vivecraft.client.ClientVRPlayers;
 import org.vivecraft.client.api_impl.VRClientAPIImpl;
 import org.vivecraft.common.api_impl.data.VRPoseHistoryImpl;
-import org.vivecraft.server.ServerNetworking;
 import org.vivecraft.server.ServerVRPlayers;
 
 import javax.annotation.Nullable;
@@ -82,15 +80,5 @@ public final class VRAPIImpl implements VRAPI {
 
     private Map<UUID, VRPoseHistoryImpl> getMap(boolean isClientSide) {
         return isClientSide ? this.clientPoseHistories : this.serverPoseHistories;
-    }
-
-    @Override
-    public void triggerHapticPulse(
-        ServerPlayer player, VRBodyPart bodyPart, float duration, float frequency, float amplitude, float delay)
-    {
-        if (isVRPlayer(player)) {
-            ServerNetworking.sendHapticToClient(player, bodyPart, duration, frequency,
-                amplitude, delay);
-        }
     }
 }
